Move conversion into data provider
