--- XRTK.Ultraleap/Packages/com.xrtk.ultraleap/Runtime/Utilities.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 901c47de832e85547a55ef769cc8a58e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
