--- XRTK.Ultraleap/Packages/com.xrtk.ultraleap/Runtime/Utilities/UltraleapHandDataConverter.cs
@@ -1,218 +1,0 @@
-﻿// Copyright (c) XRTK. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-using System;
-using UnityEngine;
-using Leap;
-using XRTK.Definitions.Controllers.Hands;
-using XRTK.Definitions.Devices;
-using XRTK.Definitions.Utilities;
-using XRTK.Ultraleap.Extensions;
-using XRTK.Services;
-using XRTK.Ultraleap.Providers.Controllers;
-using XRTK.Utilities;
-using XRTK.Ultraleap.Definitions;
-
-namespace XRTK.Ultraleap.Utilities
-{
-    /// <summary>
-    /// Converts Ultraleap hand data to <see cref="HandData"/>.
-    /// </summary>
-    public sealed class UltraleapHandDataConverter
-    {
-        /// <summary>
-        /// Constructor.
-        /// </summary>
-        /// <param name="dataProvider">The <see cref="UltraleapHandControllerDataProvider"/> using this converter.</param>
-        public UltraleapHandDataConverter(UltraleapHandControllerDataProvider dataProvider)
-        {
-            this.dataProvider = dataProvider;
-        }
-
-        private readonly UltraleapHandControllerDataProvider dataProvider;
-        private readonly MixedRealityPose[] jointPoses = new MixedRealityPose[HandData.JointCount];
-
-        private const int leapThumbIndex = (int)Finger.FingerType.TYPE_THUMB;
-        private const int leapIndexFingerIndex = (int)Finger.FingerType.TYPE_INDEX;
-        private const int leapMiddleFingerIndex = (int)Finger.FingerType.TYPE_MIDDLE;
-        private const int leapRingFingerIndex = (int)Finger.FingerType.TYPE_RING;
-        private const int leapLittleFingerIndex = (int)Finger.FingerType.TYPE_PINKY;
-
-        private const int leapMetacarpalBoneIndex = (int)Bone.BoneType.TYPE_METACARPAL;
-        private const int leapProximalBoneIndex = (int)Bone.BoneType.TYPE_PROXIMAL;
-        private const int leapIntermediateBoneIndex = (int)Bone.BoneType.TYPE_INTERMEDIATE;
-        private const int leapDistalBoneIndex = (int)Bone.BoneType.TYPE_DISTAL;
-
-        /// <summary>
-        /// Reads hand APIs for the current frame and converts it to agnostic <see cref="HandData"/>.
-        /// </summary>
-        /// <param name="handedness">The handedness of the hand to get <see cref="HandData"/> for.</param>
-        /// <param name="handData">The output <see cref="HandData"/>.</param>
-        /// <returns>True, if data conversion was a success.</returns>
-        public bool TryGetHandData(Hand hand, out HandData handData)
-        {
-            if (hand == null)
-            {
-                handData = default;
-                return false;
-            }
-
-            handData = new HandData
-            {
-                TrackingState = TrackingState.Tracked,
-                UpdatedAt = DateTimeOffset.UtcNow.Ticks
-            };
-
-            if (handData.TrackingState == TrackingState.Tracked)
-            {
-                var offsetPose = GetHandOffsetPose(dataProvider.OperationMode, dataProvider.LeapControllerOffset);
-                var handRootPose = GetHandRootPose(hand);
-
-                handData.RootPose = offsetPose + handRootPose;
-                handData.Joints = GetJointPoses(hand, handRootPose, dataProvider.OperationMode, dataProvider.LeapControllerOffset);
-                handData.Mesh = new HandMeshData();
-            }
-
-            return true;
-        }
-
-        private MixedRealityPose[] GetJointPoses(Hand hand, MixedRealityPose handRootPose, UltraleapOperationMode operationMode, Vector3 offset)
-        {
-            for (var i = 0; i < HandData.JointCount; i++)
-            {
-                var trackedHandJoint = (TrackedHandJoint)i;
-                var position = Vector3.zero;
-                var rotation = Quaternion.identity;
-
-                switch (trackedHandJoint)
-                {
-                    case TrackedHandJoint.Wrist:
-                        position = hand.WristPosition.ToLeftHandedUnityVector3();
-                        rotation = hand.Arm.Basis.rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.Palm:
-                        position = hand.PalmPosition.ToLeftHandedUnityVector3();
-                        rotation = hand.Basis.rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.ThumbMetacarpal:
-                        position = hand.Fingers[leapThumbIndex].bones[leapMetacarpalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapThumbIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.ThumbProximal:
-                        position = hand.Fingers[leapThumbIndex].bones[leapProximalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapThumbIndex].bones[leapIntermediateBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.ThumbDistal:
-                        position = hand.Fingers[leapThumbIndex].bones[leapIntermediateBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapThumbIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.ThumbTip:
-                        position = hand.Fingers[leapThumbIndex].bones[leapDistalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapThumbIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.IndexProximal:
-                        position = hand.Fingers[leapIndexFingerIndex].bones[leapMetacarpalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapIndexFingerIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.IndexIntermediate:
-                        position = hand.Fingers[leapIndexFingerIndex].bones[leapProximalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapIndexFingerIndex].bones[leapIntermediateBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.IndexDistal:
-                        position = hand.Fingers[leapIndexFingerIndex].bones[leapIntermediateBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapIndexFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.IndexTip:
-                        position = hand.Fingers[leapIndexFingerIndex].bones[leapDistalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapIndexFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.MiddleProximal:
-                        position = hand.Fingers[leapMiddleFingerIndex].bones[leapMetacarpalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapMiddleFingerIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.MiddleIntermediate:
-                        position = hand.Fingers[leapMiddleFingerIndex].bones[leapProximalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapMiddleFingerIndex].bones[leapIntermediateBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.MiddleDistal:
-                        position = hand.Fingers[leapMiddleFingerIndex].bones[leapIntermediateBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapMiddleFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.MiddleTip:
-                        position = hand.Fingers[leapMiddleFingerIndex].bones[leapDistalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapMiddleFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.RingProximal:
-                        position = hand.Fingers[leapRingFingerIndex].bones[leapMetacarpalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapRingFingerIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.RingIntermediate:
-                        position = hand.Fingers[leapRingFingerIndex].bones[leapProximalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapRingFingerIndex].bones[leapIntermediateBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.RingDistal:
-                        position = hand.Fingers[leapRingFingerIndex].bones[leapIntermediateBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapRingFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.RingTip:
-                        position = hand.Fingers[leapRingFingerIndex].bones[leapDistalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapRingFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.LittleMetacarpal:
-                        position = hand.Fingers[leapLittleFingerIndex].bones[leapMetacarpalBoneIndex].PrevJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapLittleFingerIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.LittleProximal:
-                        position = hand.Fingers[leapLittleFingerIndex].bones[leapMetacarpalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapLittleFingerIndex].bones[leapProximalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.LittleIntermediate:
-                        position = hand.Fingers[leapLittleFingerIndex].bones[leapProximalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapLittleFingerIndex].bones[leapIntermediateBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.LittleDistal:
-                        position = hand.Fingers[leapLittleFingerIndex].bones[leapIntermediateBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapLittleFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                    case TrackedHandJoint.LittleTip:
-                        position = hand.Fingers[leapLittleFingerIndex].bones[leapDistalBoneIndex].NextJoint.ToLeftHandedUnityVector3();
-                        rotation = hand.Fingers[leapLittleFingerIndex].bones[leapDistalBoneIndex].Rotation.ToLeftHandedUnityQuaternion();
-                        break;
-                }
-
-                jointPoses[i] = new MixedRealityPose(position - handRootPose.Position, rotation);
-            }
-
-            // Fill missing joints by estimating their pose.
-            jointPoses[(int)TrackedHandJoint.RingMetacarpal] = HandUtilities.GetEstimatedRingMetacarpalPose(jointPoses);
-            jointPoses[(int)TrackedHandJoint.MiddleMetacarpal] = HandUtilities.GetEstimatedMiddleMetacarpalPose(jointPoses);
-            jointPoses[(int)TrackedHandJoint.IndexMetacarpal] = HandUtilities.GetEstimatedIndexMetacarpalPose(jointPoses);
-
-            return jointPoses;
-        }
-
-        private MixedRealityPose GetHandRootPose(Hand hand)
-        {
-            var position = hand.WristPosition.ToLeftHandedUnityVector3();
-            var rotation = hand.Arm.Basis.rotation.ToLeftHandedUnityQuaternion();
-
-            return new MixedRealityPose(position, rotation);
-        }
-
-        private MixedRealityPose GetHandOffsetPose(UltraleapOperationMode operationMode, Vector3 offset)
-        {
-            switch (operationMode)
-            {
-                case UltraleapOperationMode.Desktop:
-                    var cameraTransform = MixedRealityToolkit.CameraSystem != null
-                        ? MixedRealityToolkit.CameraSystem.MainCameraRig.PlayerCamera.transform
-                        : CameraCache.Main.transform;
-
-                    return new MixedRealityPose(cameraTransform.position + offset, Quaternion.identity);
-                case UltraleapOperationMode.HeadsetMounted:
-                default:
-                    return MixedRealityPose.ZeroIdentity;
-            }
-        }
-    }
-}
