--- XRTK.Ultraleap/Packages/com.xrtk.ultraleap/Runtime/Utilities/UltraleapHandDataConverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 008033fdb87c1004b95328794a80945f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
